post date migration

+++ src/routes/web.php
@@ -3,6 +3,7 @@
 use API\Platform\Models\Table;
 use API\Platform\Services\ORMService;
 use Illuminate\Support\Facades\Route;
+use Illuminate\Support\Facades\Schema;
 
 # @todo middleware check dev to expose
 
@@ -11,13 +12,16 @@ Route::post('schema', '\Agontuk\Schema\Controllers\SchemaController@generateMigr
 
 $ormService = new ORMService();
 
-$tables = Table::all();
-foreach ($tables as $table) {
-    $modelName = $ormService->tableToClass($table, $table->name);
+// check given migrations might not be done
+if (Schema::connection('mysql')->hasTable('tables')) {
+    $tables = Table::all();
+    foreach ($tables as $table) {
+        $modelName = $ormService->tableToClass($table, $table->name);
 
-    Route::get('api/' . $table->name, '\App\Http\Controllers\API\\' . $modelName . 'APIController@index');
-    Route::post('api/' . $table->name, '\App\Http\Controllers\API\\'. $modelName . 'APIController@store');
-    Route::get('api/' . $table->name . '/{id}', '\App\Http\Controllers\API\\' . $modelName . 'APIController@show');
-    Route::put('api/' . $table->name . '/{id}', '\App\Http\Controllers\API\\' . $modelName . 'APIController@update');
-    Route::delete('api/' . $table->name . '/{id}', '\App\Http\Controllers\API\\' . $modelName . 'APIController@destroy');
+        Route::get('api/' . $table->name, '\App\Http\Controllers\API\\' . $modelName . 'APIController@index');
+        Route::post('api/' . $table->name, '\App\Http\Controllers\API\\'. $modelName . 'APIController@store');
+        Route::get('api/' . $table->name . '/{id}', '\App\Http\Controllers\API\\' . $modelName . 'APIController@show');
+        Route::put('api/' . $table->name . '/{id}', '\App\Http\Controllers\API\\' . $modelName . 'APIController@update');
+        Route::delete('api/' . $table->name . '/{id}', '\App\Http\Controllers\API\\' . $modelName . 'APIController@destroy');
+    }
 }
